+++ src/index.ts
@@ -2,4 +2,5 @@
 
 export * from "./md5";
 export * from "./parse";
+export * from "./parseRos2idl";
 export * from "./stringify";
